requirements.txt
numpy==2.4.3
omegaconf==2.3.0
Pillow==12.1.1
scikit_learn==1.8.0
torch==2.10.0
torchvision==0.25.0
setuptools==80.10.2
tensorboard==2.20.0
